Configurações de ambiente concluídas

+++ src/main/java/br/edu/thejukebox/repository/AccountRepository.java
@@ -4,6 +4,6 @@ package br.edu.thejukebox.repository;
 import br.edu.thejukebox.model.Account;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AccountRepository extends CrudRepository<Account, Integer> {
+public interface AccountRepository extends CrudRepository<Account, Long> {
     Account findAccountByUsername(String username);
 }
