+++ src/main/java/br/edu/thejukebox/repository/AlbumRepository.java
@@ -3,6 +3,6 @@ package br.edu.thejukebox.repository;
 import br.edu.thejukebox.model.Album;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AlbumRepository extends CrudRepository<Album, Integer> {
+public interface AlbumRepository extends CrudRepository<Album, Long> {
 
 }
